Struct apache_log_security_lib::config::Config

pub struct Config {
    pub reporting: Vec<Reporting>,
    pub services: Vec<Service>,
    pub xss_level: XssLevel,
}

Config struct required to run the library

Fields
reporting: Vec<Reporting>

Where to report incidents

services: Vec<Service>

Sources for logs

xss_level: XssLevel
Methods

impl Config

pub fn new() -> Config

Create default configuration. With no inputs and outputs. Is used for generating default config file.

pub fn normalize_services_glob(self) -> Config

Is used to convert glob paths to concrete paths.

If we have directory dir with files a.log and b.log this will turn dir/* to concrete paths dir/a.log and dir/b.log and return new config while consuming the old one.