Function apache_log_security_lib::run

pub fn run(conf: Config) -> Result<(), LibErr>

Run library with given configuration. Library is divided into three layers. Each layers contains modules that can be configued to perform required task be specifing their composition using Config structure. Logs are beeing read, analyzed and then reported as configured.

Runing with default configuration:

use apache_log_security_lib::config::Config;

let config = Config::new();
run(config);

Runing with custom configuration. Using non-verbose std output and basic Xss detection.

use apache_log_security_lib::config::Config;
use apache_log_security_lib::config::XssLevel;
use apache_log_security_lib::config::Reporting;
use apache_log_security_lib::reporting::std::Std;

let config = Config {
   reporting: vec![Reporting::Std(Std{verbose: false})],
   services: vec![],
   xss_level: XssLevel::Basic,
};
run(config);